Refactors how Actions are created
Co-Authored_By: Alex Gonzalez <agonzalez@codeforamerica.org>
